document apdb library functions
git-svn-id: https://jff.googlecode.com/svn/trunk@85 1de5394f-e846-0410-8a38-9b14b2994ffe

+++ abbs/board.c
@@ -116,7 +116,7 @@ board_open(const char* path, char mode)
             goto L_error;
 
         memset(&index, 0, sizeof(index_t));
-        if (-1 == apdb_add_begin(board->db, sizeof(h)) ||
+        if ((unsigned)-1 == apdb_add_begin(board->db, sizeof(h)) ||
                 -1 == apdb_append_data(board->db, &h, sizeof(h)) ||
                 -1 == apdb_add_end(board->db, &index))
             goto L_error;
